refactor test helpers into jest setup
