+++ detox/local-cli/build-framework-cache.test.js
@@ -1,18 +1,8 @@
-const yargs = require('yargs');
-
-function call(cmd) {
-  const parser = yargs
-    .scriptName('detox')
-    .command(require('./build-framework-cache'))
-    .help();
-  return new Promise(resolve => parser.parse(cmd, (_err, _argv, output) => resolve(output)));
-}
-
 describe('build-framework-cache', () => {
   it('shows help text', async () => {
     jest.spyOn(process, 'exit'); // otherwise tests are aborted
 
-    expect(await call('--help')).toMatchInlineSnapshot(`
+    expect(await callCli('./build-framework-cache', '--help')).toMatchInlineSnapshot(`
 "detox [command]
 
 Commands:
@@ -29,7 +19,7 @@ Options:
   it('executes shell script', async () => {
     jest.mock('child_process');
     const cp = require('child_process');
-    await call('build-framework-cache');
+    await callCli('./build-framework-cache', 'build-framework-cache');
 
     expect(cp.execSync).toHaveBeenCalledWith(expect.stringContaining('scripts/build_framework.ios.sh'), expect.any(Object));
   });
