+++ detox/local-cli/clean-framework-cache.test.js
@@ -1,18 +1,8 @@
-const yargs = require('yargs');
-
-function call(cmd) {
-  const parser = yargs
-    .scriptName('detox')
-    .command(require('./clean-framework-cache'))
-    .help();
-  return new Promise(resolve => parser.parse(cmd, (_err, _argv, output) => resolve(output)));
-}
-
 describe('clean-framework-cache', () => {
   it('shows help text', async () => {
     jest.spyOn(process, 'exit'); // otherwise tests are aborted
 
-    expect(await call('--help')).toMatchInlineSnapshot(`
+    expect(await callCli('./clean-framework-cache', '--help')).toMatchInlineSnapshot(`
 "detox [command]
 
 Commands:
@@ -36,7 +26,7 @@ Options:
       remove: jest.fn()
     }));
     const fs = require('fs-extra');
-    await call('clean-framework-cache');
+    await callCli('./clean-framework-cache', 'clean-framework-cache');
 
     expect(fs.remove).toHaveBeenCalledWith('/home/Library/Detox');
   });
@@ -51,7 +41,7 @@ Options:
       remove: jest.fn()
     }));
     const fs = require('fs-extra');
-    await call('clean-framework-cache');
+    await callCli('./clean-framework-cache', 'clean-framework-cache');
 
     expect(fs.remove).not.toHaveBeenCalled();
   });
